authentication/requirements.txt
fastapi
uvicorn
python-multipart
python-jose[cryptography]
passlib[bcrypt]
sqlalchemy
python-dotenv
sqlmodel
psycopg2-binary
alembic
